+++ build-transforms.py
@@ -5,6 +5,35 @@ import sys
 import os
 from config import *
 
+VERSION = "2.0"
+
+
+def row_to_itds_row(row, transform_function_name):
+    description = row["description"]
+    ui_name = description.split(":")[1].strip()
+    ui_name = f"{ui_name} [OpenCTI]"
+    input_type = row["input_type"]
+    host = trx_server_host
+    if host.endswith("/"):
+        host = host[:-1]
+
+    return {
+        "Owner": "ANSSI",
+        "Author": "Maltego, ANSSI",
+        "Disclaimer": "",
+        "Description": description,
+        "Version": VERSION,
+        "Name": transform_function_name,
+        "UIName": ui_name,
+        "URL": f"{host}/run/{transform_function_name}",
+        "entityName": input_type,
+        "oAuthSettingId": "",
+        "transformSettingIDs": "OpenCTIURL,OpenCTIToken,SSLVerify,HTTPProxies",
+        "seedIDs": itds_seed_name,
+    }
+
+
+
 input_name = sys.argv[1]
 keys = ["Transform id", "Description", "Input type", "Sets", "Output"]
 
@@ -24,6 +53,10 @@ set_template = ""
 with open("./templates/template.set", "r") as i:
     set_template = i.read()
 
+python_template_file = ""
+with open("./templates/transform.py", "r") as i:
+    python_template_file = i.read()
+
 csv_content = []
 with open(input_name) as csvfile:
     r = csv.reader(csvfile, delimiter=",", quotechar="'")
@@ -32,6 +65,7 @@ with open(input_name) as csvfile:
 
 transforms = ""
 transforms_sets = {}
+itds_transforms_rows = []
 
 for row in csv_content[1:]:
     d = {}
@@ -61,15 +95,15 @@ for row in csv_content[1:]:
                 transforms_sets[cset] = []
             transforms_sets[cset].append(row[d["Transform id"]])
 
-    options = ""
-    if row[d["Output"]] != "":
-        options = " --output " + row[d["Output"]]
+    transform_function = row[d["Transform id"]].replace(".","").replace("-","").replace("_","")
+
+    itds_row = row_to_itds_row(t_data, transform_function)
+    itds_transforms_rows.append(itds_row)
 
     ts_data = {
         "python_path": python_path,
         "local_execution_path": local_execution_path,
-        "transform_name": row[d["Transform id"]].split(".")[-1],
-        "options": options,
+        "transform_function": transform_function,
     }
     ts_output = ts_template.format(**ts_data)
     with open(
@@ -81,6 +115,13 @@ for row in csv_content[1:]:
         o.write(ts_output)
 
     transforms += '      <Transform name="' + row[d["Transform id"]] + '"/>\n'
+
+    with open("./trx/gunicorn/transforms/" + transform_function + ".py", "w") as o:
+        o.write(python_template_file.format(
+            functionName=transform_function,
+            transformName=row[d["Transform id"]].split(".")[-1],
+            output=('"' + row[d["Output"]] + '"') if row[d["Output"]] != "" else 'None'
+        ))
 
 s_data = {"transforms": transforms[:-1]}
 s_output = s_template.format(**s_data)
@@ -102,3 +143,28 @@ for cset, ctransforms in transforms_sets.items():
         "w",
     ) as o:
         o.write(set_output)
+
+    tds_set_data = {
+        "transformSetName": cset,
+        "transforms": "\n".join(
+            list(map(lambda r: '      <Transform name="paterva.v2.' + r.replace(".","").replace("-","").replace("_","") + '"/>', ctransforms))
+        ),
+    }
+    tds_set_output = set_template.format(**tds_set_data)
+
+    with open(
+        "./mtz/TransformSetsTDS/"
+        + cset.lower().replace(" ", "").replace(":", "")
+        + ".set",
+        "w",
+    ) as o:
+        o.write(tds_set_output)
+
+if itds_transforms_rows:
+    with open("output/importable_itds_config.csv", "w") as outf:
+        writer = csv.DictWriter(
+            outf, fieldnames=list(itds_transforms_rows[0].keys()), delimiter=";", quoting=csv.QUOTE_ALL
+        )
+        writer.writeheader()
+        for data in itds_transforms_rows:
+            writer.writerow(data)