+++ build_transforms.sh
@@ -2,16 +2,18 @@
 
 echo "Clear old results"
 rm output/transforms.mtz 2> /dev/null
-rm -R mtz/Servers/ mtz/TransformSets/ mtz/TransformRepositories/ 2> /dev/null
+find ./trx/gunicorn/transforms -name 'opencti*' -exec rm -f {} \; 2> /dev/null
+rm -R mtz/Servers/ mtz/TransformSets/ mtz/TransformSetsTDS/ mtz/TransformRepositories/ 2> /dev/null
 mkdir mtz/ 2> /dev/null
 mkdir mtz/Servers/
 mkdir mtz/TransformSets/
+mkdir mtz/TransformSetsTDS/
 mkdir mtz/TransformRepositories/
 mkdir mtz/TransformRepositories/Local/
 mkdir output/ 2> /dev/null
 
 echo "Generate Maltego transforms config"
-python3 generate_transforms.py transforms.csv
+python3 build-transforms.py transforms.csv
 
 echo "Create MTZ package"
 cd mtz
@@ -19,7 +21,7 @@ zip -q -x .empty -r ../output/transforms.mtz ./Servers ./TransformRepositories .
 cd ../
 
 # Add opencti configuration file next to transforms
-echo "Copy config.py in ./src"
-cp config.py ./src/config.py
+echo "Copy config.py in ./trx/gunicorn/opencti/"
+cp config.py ./trx/gunicorn/opencti/config.py
 
-echo "All done. MTZ packages can be imported in Maltego and don't forget to copy ./src/ at the execution path provided in the configuration."
+echo "All done. MTZ packages can be imported in Maltego and don't forget to copy ./trx/gunicorn/ at the execution path provided in the configuration."