+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## Version 1.0 (14/04/2020)
 
 Release of a first stable version
+
+## Version 2.0 (30/03/2020)
+
+Release of a new version fully compatible with OpenCTI v4.3 and the revamped STIX2.1 Maltego entities available in the official Maltego Transform Hub.
+Depends on maltego-trx and maltego-stix2 python packages.